+++ Patches/RoundManagerPatch.cs
@@ -69,5 +69,18 @@ namespace GeneralImprovements.Patches
                 _gotShipNode = false;
             }
         }
+
+        [HarmonyPatch(typeof(RoundManager), "RefreshEnemiesList")]
+        [HarmonyPatch(typeof(EnemyAI), "SubtractFromPowerLevel")]
+        [HarmonyPatch(typeof(RoundManager), "AdvanceHourAndSpawnNewBatchOfEnemies")]
+        [HarmonyPatch(typeof(RoundManager), "DespawnEnemyGameObject")]
+        [HarmonyPatch(typeof(RoundManager), "UnloadSceneObjectsEarly")]
+        [HarmonyPostfix]
+        private static void RoundLevelPowerChanged(RoundManager __instance)
+        {
+            float totalMaxPower = __instance.currentMaxInsidePower + __instance.currentMaxOutsidePower;
+            float totalCurrentPower = __instance.currentEnemyPower + __instance.currentOutsideEnemyPower;
+            MonitorsHelper.UpdateDangerLevelMonitors(totalMaxPower, totalCurrentPower);
+        }
     }
 }